aiart-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
python-multipart = "^0.0.18"
Pillow = "^11.0.0"
ffmpeg-python = "^0.2.0"
imageio-ffmpeg = "^0.5.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
slowapi = "^0.1.9"
motor = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
